+++ packages/algob/src/internal/deployer.ts
@@ -6,7 +6,7 @@ import { cloneDeep } from "lodash";
 
 import { txWriter } from "../internal/tx-log-writer";
 import { AlgoOperator } from "../lib/algo-operator";
-import { getBytecodeAndHash } from "../lib/compile";
+import { CompileOp, getBytecodeAndHash } from "../lib/compile";
 import { getDummyLsig, getLsig, getLsigFromCache } from "../lib/lsig";
 import { blsigExt, loadBinaryLsig, readMsigFromFile } from "../lib/msig";
 import { CheckpointFunctionsImpl, persistCheckpoint } from "../lib/script-checkpoints";
@@ -196,14 +196,22 @@ class DeployerBasicMode {
   }
 
   /**
-   * Returns program compiled info: bytecode (compiled program) and it's hash.
+   * Complies program in real time, and returns compiled info (bytecode, hash, filename etc).
    * @param name ASC name
    * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
-   * @returns loaded logic signature from artifacts/cache/<file_name>.teal.yaml
    */
   async getCompiledASC (name: string,
     scTmplParams?: SCParams): Promise<ASCCache> {
     return await getBytecodeAndHash(name, this.algoOp.algodClient, scTmplParams);
+  }
+
+  /**
+   * Returns cached program (in artifacts/cache) compiled info(bytecode, hash, filename etc).
+   * @param name ASC file name
+   */
+  async getDeployedASC (name: string): Promise<ASCCache | undefined> {
+    const op = new CompileOp(this.algoOp.algodClient);
+    return await op.readArtifact(name);
   }
 
   /**