+++ packages/algob/src/types.ts
@@ -665,9 +665,14 @@ export interface Deployer {
    * Returns program compiled info: bytecode (compiled program) and it's hash.
    * @param name ASC name
    * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
-   * @returns loaded logic signature from artifacts/cache/<file_name>.teal.yaml
    */
   getCompiledASC: (name: string, scTmplParams?: SCParams) => Promise<ASCCache>
+
+  /**
+   * Returns cached program (in artifacts/cache) compiled info(bytecode, hash, filename etc).
+   * @param name ASC file name
+   */
+  getDeployedASC: (name: string) => Promise<ASCCache | undefined>
 
   /**
    * Checks if checkpoint is deleted for a particular transaction