+++ packages/algob/test/mocks/deployer.ts
@@ -126,6 +126,10 @@ export class FakeDeployer implements Deployer {
     throw new Error("Not implemented");
   }
 
+  async getDeployedASC (name: string): Promise<ASCCache | undefined> {
+    throw new Error("Not implemented");
+  }
+
   async deployASADef (
     name: string,
     asaDef: wtypes.ASADef,